+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 - Specify Calculation-Energy and Geometry-Energy as many-to-one relationships
+- Uniqueness constraint in CalculationHashRow
 
 ## [0.0.2] - 2026-01-28
 ### Added